% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LML.R
\name{LMc}
\alias{LMc}
\title{LMc}
\usage{
LMc(new_el, old_list = NULL, M = 1)
}
\arguments{
\item{new_el}{elements to add to top of list, if length(new_el) > 1, then the new elements will be appended left to right (so that the last element will be the most recent).}

\item{old_list}{elements to concatenate new_el with}

\item{M}{maximum number of elements (will truncate by removing oldest if length is greater than M)}
}
\description{
Limited Memory concatenate. Adds new elements to front, removing oldest elements if new length is greater tham M
}
\examples{
a = LMc(1, M=5)
print(a)

a = LMc(c(2,3,4,5), a, M=5)
print(a)

a = LMc(c(6,7), a, M=5)
print(a)
}
