% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LML.R
\name{LMList}
\alias{LMList}
\title{LMList}
\usage{
LMList(new_el, old_list = list(), M = 1)
}
\arguments{
\item{new_el}{element to add to top of list}

\item{old_list}{list to append new_el to}

\item{M}{maximum size of list (will truncate by removing bottom if length is greater than M)}
}
\description{
Limited Memory List. Adds new element to top of list, removing bottom elements if new length is greater tham M.
}
\examples{
# create a new LM List, with 2 element memory:
lm = LMList("horse", M=2)
print(lm)

# append new item
lm = LMList("buggy", lm, M=2)
print(lm)

# append a third item (will forget the oldest)
lm = LMList("driver", lm, M=2)
print(lm)
}
